Moving ReposeJettyServer into a new package
