+++ repose-aggregator/artifacts/valve/src/main/scala/org/openrepose/valve/jetty/ReposeJettyServer.scala
@@ -17,7 +17,7 @@
  * limitations under the License.
  * =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
  */
-package org.openrepose.valve
+package org.openrepose.valve.jetty
 
 import java.util
