+++ repose-aggregator/artifacts/valve/src/main/scala/org/openrepose/valve/jmx/ValvePortMXBeanImpl.scala
@@ -20,11 +20,10 @@
 package org.openrepose.valve.jmx
 
 import java.util.concurrent.ConcurrentHashMap
+
 import javax.inject.Named
-
-import org.openrepose.valve.ReposeJettyServer
+import org.openrepose.valve.jetty.ReposeJettyServer
 import org.springframework.jmx.export.annotation.{ManagedAttribute, ManagedResource}
-
 
 @Named
 @ManagedResource(
